Code cleanup and fixes (#382)
* Rework Soulbound mechanic

Rework Soulbound mechanic to prevent a memory leak if a player disconnects before respawning.

* Move namespaced key location

* Add mending support for the durability mechanic

* Fix customblockdata not being relocated

* Remove redundant snakeyaml from plugin jar

* Add missing relocations

* Update Gradle and Shadow

* Fix project imports

* Code cleanup

* Use apache http for resourcepack instead of java-curl

* Remove * imports

* Bug fixes

Fix error with ItemBuilder and revert changes to readFileToVirtuals

+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/gameplay/noteblock/NoteBlockMechanicListener.java
@@ -355,7 +355,6 @@ public class NoteBlockMechanicListener implements Listener {
         Bukkit.getPluginManager().callEvent(blockPlaceEvent);
 
         if (!BlockHelpers.correctAllBlockStates(target, player, face, item)) blockPlaceEvent.setCancelled(true);
-
         if (!blockPlaceEvent.canBuild() || blockPlaceEvent.isCancelled()) {
             target.setBlockData(curentBlockData, false); // false to cancel physic
             return null;
@@ -381,7 +380,7 @@ public class NoteBlockMechanicListener implements Listener {
         if (block.getType() != Material.NOTE_BLOCK) return null;
         final NoteBlock noteBlok = (NoteBlock) block.getBlockData();
         return NoteBlockMechanicFactory
-                .getBlockMechanic((int) (noteBlok.getInstrument().getType()) * 25
-                        + (int) noteBlok.getNote().getId() + (noteBlok.isPowered() ? 400 : 0) - 26);
+                .getBlockMechanic((noteBlok.getInstrument().getType()) * 25
+                        + noteBlok.getNote().getId() + (noteBlok.isPowered() ? 400 : 0) - 26);
     }
 }
